"Refactor CompactorDeletorPreview logic for better structure"
Split the logic for CompactorDeletorPreview to reduce the complexity within the HandledScreenMixin.java file. This change helps to make the code simpler and easier to maintain.
